requirements.txt
Django==1.11.5
django-annoying==0.10.3
olefile==0.44
Pillow==4.2.1
pytz==2017.2
